% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig}
\alias{Fig}
\title{The curves of Cumulative Cause-specific Hazard or Cumulative Incidence Function.}
\usage{
Fig(time, status, group, event.type = 1, fig = 1,
           x.lim = c(0, 10), y.lim = c(0, 1),
           x.space = seq(0, 10, 2), y.space = seq(0, 1, 0.2),
           group.lab = c("group 0", "group 1"),
           col = c("steelblue", "red"), lty = c(1, 2))
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1 = event of interest, 2 = competing event and 0 = right censored.}

\item{group}{The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0 = control group, 1 = active treatment group.}

\item{event.type}{The indicator of event type, 1 = event 1, 2 = event 2.}

\item{fig}{The indicator of curves, 1 = Cumulative incidence function, 2 = Cumulative Cause-specific Hazard.}

\item{x.lim}{The range of x-axis.}

\item{y.lim}{The range of y-axis.}

\item{x.space}{The scale of x-axis.}

\item{y.space}{The scale of y-axis.}

\item{group.lab}{The labels of groups.}

\item{col}{The color of curves.}

\item{lty}{The type of curves.}
}
\value{
The curves of Cumulative Cause-specific Hazard or Cumulative Incidence Function.
}
\description{
Drawing the curves of Cumulative Cause-specific Hazard or Cumulative Incidence Function.
}
\examples{
data(simdata)
time <- simdata$time
status <- simdata$status
group <- simdata$group
open <- par(no.readonly = TRUE)
par(mfrow = c(2, 2))
Fig(time, status, group, event.type = 1, fig = 1,
   x.lim = c(0, 6), y.lim = c(0, 0.8),
   group.lab = c("RT", "CRT"), col = c("steelblue", "red"), lty = c(1, 2))
Fig(time, status, group, event.type = 2, fig = 1,
   x.lim = c(0, 6), y.lim = c(0, 0.8),
   group.lab = c("RT", "CRT"), col = c("steelblue", "red"), lty = c(1, 2))
Fig(time, status, group, event.type = 1, fig = 2,
   x.lim = c(0, 6), y.lim = c(0, 0.8),
   group.lab = c("RT", "CRT"), col = c("steelblue", "red"), lty = c(1, 2))
Fig(time, status, group, event.type = 2, fig = 2,
   x.lim = c(0, 6), y.lim = c(0, 0.8),
   group.lab = c("RT", "CRT"), col = c("steelblue", "red"), lty = c(1, 2))
par(open)
}
