% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMTLd_sample.R
\name{RMTLd_sample}
\alias{RMTLd_sample}
\title{Sample size based on the difference in restricted mean time lost (RMTLd)}
\usage{
RMTLd_sample(alpha, beta, r, delta, var.0, var.1)
}
\arguments{
\item{alpha}{The default is 0.05.}

\item{beta}{The default is 0.20.}

\item{r}{The ratio (group 1 over group 0) of the sample size between groups}

\item{delta}{A value to specify RMTLd.}

\item{var.0}{A value to specify the variance in group 0.}

\item{var.1}{A value to specify the variance in group 1.}
}
\value{
an object of class RMTLd_sample
}
\description{
calculating sample size based on RMTLd
}
\examples{
library(crRMTL)
data(Example1)
test <- RMTLd_test(Example1$time, Example1$status, Example1$group, alpha = 0.05,
digits = 3, tau = NULL)
n0 <- table(Example1$group)[1]
n1 <- table(Example1$group)[2]
r <- n1 / n0
delta <- test$Test.event1[3, 1]  ## RMTL of event 1
var.0 <- test$RMTL[1, 3] * n0
var.1 <- test$RMTL[2, 3] * n1
sample <- RMTLd_sample(alpha = 0.05, beta = 0.2, r = r, delta = delta, var.0 = var.0, var.1 = var.1)
sample
}
