% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMTLd_test.R
\name{RMTLd_test}
\alias{RMTLd_test}
\title{Comparing the difference in restricted mean time lost (RMTLd) under competing risks}
\usage{
RMTLd_test(time, status, group, alpha = 0.05, digits = 3, tau = NULL)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1 = event of interest, 2 = competing
events and 0 = right censored.}

\item{group}{The group indicator for comparison. The elements of this vector
take either 1 or 0. Normally, 0 = control group, 1 = treatment group.}

\item{alpha}{The default is 0.05. (1-\code{alpha}) confidence intervals are
reported.}

\item{digits}{The decimal places settings, digits = 3 are set by default.}

\item{tau}{The value to specify the truncation time point for the RMTL
calculation.
The default is the minimum of the largest follow-up time of two groups.}
}
\value{
an object of class RMTLd_test.

\item{Note}{a note regarding the specification of the truncation
time (tau), the default tau will be reported if tau was not specified.}

\item{RMTL}{RMTL results in two groups, including RMTL of event 1
and event 2, and corresponding variance and confidence interval in group 0 and 1.}

\item{Test.event1}{
Three kinds of effect size between groups of event 1 (event of interest), including
cause-specific hazard ratio (cHR, group 1 over group 0), subdistributional Hazard Ratio
(sHR, group 1 over group 0) and RMTLd (group 1 minus group 0), and corresponding
confidence intervals. Meanwhile, the Z-statistics and P-value are estimated via log-rank test,
Gray's test and RMTLd test.}

\item{Test.event2}{
Three kinds of effect size between groups of event 2 (competing event), including
cause-specific hazard ratio (cHR, group 1 over group 0), subdistributional Hazard Ratio
(sHR, group 1 over group 0) and RMTLd (group 1 minus group 0), and corresponding
confidence intervals. Meanwhile, the Z-statistics and P-value are estimated via log-rank test,
Gray's test and RMTLd test.}
}
\description{
Performs two-sample comparisons using the difference in restricted mean time
lost (RMTLd) as a summary measure of between-group difference under
competing risks. At the same time, the outcome of hazard-based measures are
also reported for assessing between-group difference.
}
\examples{
library(crRMTL)
data(Example1)
# The time point is set as the minimum of the largest follow-up time of two groups with tau = NULL.
RMTLd_test(Example1$time, Example1$status, Example1$group, alpha = 0.05, digits = 3, tau = NULL)

# The time point can be specified as fixed time point,like tau = 10.
RMTLd_test(Example1$time, Example1$status, Example1$group, alpha = 0.05, digits = 3, tau = 10)

}
