% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{The reconstructed data of 221 patients with locally advanced
nasopharyngeal cancer in the SQNP01 trail.}
\format{
A data frame with 221 rows and 3 variables:
}
\source{
\url{https://trialsjournal.biomedcentral.com/articles/10.1186/1745-6215-12-127}
}
\usage{
data(simdata)
}
\description{
The reconstructed data of 221 patients with locally advanced
nasopharyngeal cancer in the SQNP01 trail.
}
\keyword{datasets}
